fix : cpp module 08 ex01 내부 데이터 int형으로 변환

+++ 42cursus/CPP08/ex01/Span.cpp
@@ -1,6 +1,6 @@
 #include "Span.hpp"
 
-Span::Span(int capacity): data(std::vector<unsigned int>(0)), capacity(capacity) {
+Span::Span(unsigned int capacity): data(std::vector<int>(0)), capacity(capacity) {
 
 }
 
@@ -20,7 +20,7 @@ Span& Span::operator=(const Span& other) {
 	return *this;
 }
 
-void Span::addNumber(unsigned int n) {
+void Span::addNumber(int n) {
 	if (this->data.size() == capacity)
 		throw std::logic_error("container is full");
 	this->data.push_back(n);
@@ -29,18 +29,18 @@ void Span::addNumber(unsigned int n) {
 unsigned int Span::shortestSpan() {
 	if (this->data.size() < 2)
 		throw std::logic_error("cannot found shortestSpan");
-	std::vector<unsigned int> tmp = this->data;
+	std::vector<int> tmp = this->data;
 	std::sort(tmp.begin(), tmp.end());
 	unsigned int res = ~0;
-	for (std::vector<unsigned int>::const_iterator it = tmp.begin(); it + 1 != tmp.end(); it++)
-		res = std::min(res, *(it + 1) - *it);
+	for (std::vector<int>::const_iterator it = tmp.begin(); it + 1 != tmp.end(); it++)
+		res = std::min(res, static_cast<unsigned int>(*(it + 1) - *it));
 	return res;
 }
 
 unsigned int Span::longestSpan() {
 	if (this->data.size() < 2)
 		throw std::logic_error("cannot found longestSpan");
-	std::vector<unsigned int> tmp = this->data;
+	std::vector<int> tmp = this->data;
 	std::sort(tmp.begin(), tmp.end());
-	return *(tmp.end() - 1) - *tmp.begin();
+	return static_cast<unsigned int>(*(tmp.end() - 1)) - static_cast<unsigned int>(*tmp.begin());
 }
